git-svn-id: https://enoki.dokukino.com/svn/public/fat32fs@196 0baba885-a8eb-41f5-96ae-9b54537fed86

+++ disk.h
@@ -1,5 +1,5 @@
 #ifndef DISK_H_
 #define DISK_H_
 
-size_t disk_read(int id, void *buf, off_t offset, size_t count);
+int disk_read(int id, void *buf, off_t offset, int count);
 #endif /*DISK_H_*/
